Fix issues with '~' in paths; start on rest node tests

+++ src/madsci_node_module/madsci/node_module/rest_node_module.py
@@ -4,7 +4,7 @@ import json
 import shutil
 import tempfile
 import time
-from collections.abc import Generator
+from contextlib import asynccontextmanager
 from multiprocessing import Process
 from pathlib import Path, PureWindowsPath
 from threading import Thread
@@ -119,10 +119,10 @@ class RestNode(AbstractNode):
             port=port,
         )
 
-    def start_node(self) -> None:
+    def start_node(self, testing: bool = False) -> None:
         """Start the node."""
         super().start_node()  # *Kick off protocol agnostic-startup
-        self._start_rest_api()
+        self._start_rest_api(testing=testing)
 
     """------------------------------------------------------------------------------------------------"""
     """Interface Methods"""
@@ -253,7 +253,7 @@ class RestNode(AbstractNode):
     """Internal and Private Methods"""
     """------------------------------------------------------------------------------------------------"""
 
-    def _start_rest_api(self) -> None:
+    def _start_rest_api(self, testing: bool = False) -> None:
         """Start the REST API for the node."""
         import uvicorn
 
@@ -261,22 +261,23 @@ class RestNode(AbstractNode):
         self._configure_routes()
         host = getattr(self.config, "host", "localhost")
         port = getattr(self.config, "port", 2000)
-        self.rest_server_process = Process(
-            target=uvicorn.run,
-            args=(self.rest_api,),
-            kwargs={"host": host, "port": port},
-            daemon=True,
-        )
-        self.rest_server_process.start()
-        while True:
-            time.sleep(1)
-            if self.restart_flag:
-                self.rest_server_process.terminate()
-                self.restart_flag = False
-                self._start_rest_api()
-                break
-            if self.exit_flag:
-                break
+        if not testing:
+            self.rest_server_process = Process(
+                target=uvicorn.run,
+                args=(self.rest_api,),
+                kwargs={"host": host, "port": port},
+                daemon=True,
+            )
+            self.rest_server_process.start()
+            while True:
+                time.sleep(1)
+                if self.restart_flag:
+                    self.rest_server_process.terminate()
+                    self.restart_flag = False
+                    self._start_rest_api()
+                    break
+                if self.exit_flag:
+                    break
 
     def _startup_thread(self) -> None:
         """The startup thread for the REST API."""
@@ -296,7 +297,8 @@ class RestNode(AbstractNode):
             self.logger.log(f"Startup complete for node {self.node_info.node_name}.")
             self.node_status.initializing = False
 
-    def _lifespan(self, app: FastAPI) -> Generator[None, None, None]:  # noqa: ARG002
+    @asynccontextmanager
+    async def _lifespan(self, app: FastAPI):  # noqa: ANN202, ARG002
         """The lifespan of the REST API."""
         # * Run startup on a separate thread so it doesn't block the rest server from starting
         # * (node won't accept actions until startup is complete)
